requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
Flask-CORS
whitenoise
Pillow
pypdfium2
ultralytics
torch
torchvision
numpy<2.0
opencv-python>=4.9.0
python-socketio
matplotlib
pandas
Werkzeug
roboflow
PyYAML
git+https://github.com/facebookresearch/segment-anything-2.git
shapely
scipy
huggingface_hub
